+++ mock/index.ts
@@ -1,7 +1,9 @@
 import type { MockMethod } from 'vite-plugin-mock';
 
 import login from './auth/login';
+import menuList from './auth/menuList';
 
 export default [
-  login
+  login,
+  menuList
 ] as MockMethod[];
