+++ src/api/login.ts
@@ -8,3 +8,12 @@ export function loginApi<T>(data: T) {
     data
   });
 }
+
+// 获取用户的菜单
+export function menuListApi<T>(data: T) {
+    return service({
+      url: '/mock/menuList',
+      method: 'POST',
+      data
+    });
+  }
